Combine media report tests and use factories and stricter mock call checks
